app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Bodya3836 <kotdanon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
sqlalchemy = "^2.0.25"
psycopg2 = "^2.9.9"
myapplication = "^0.1.0"
path = "^16.9.0"
alembic = "^1.13.1"
pyjwt = "^2.8.0"
fastapi-mail = "^1.4.1"
cloudinary = "^1.38.0"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
